# Improve changelog widget performance