Improve Logging::fetch_all_changelogs SQL and changelog widget load performance (3.5 to 0.20s).

Can you also index changelog.ctype when you next bump the SCHEMA version? This further improves the Logging::fetch_all_changelogs SQL query performance (0.20 to 0.01s).

ALTER TABLE `changelog` ADD INDEX(`ctype`);


[MySQL 5.1, RedHat x64]

-----------------------------------------------------

The changelog widget is taking 3.15s to render and calls
Logging::fetch_all_changelogs to display the last 5 change entries.

Running EXPLAIN on the captured Logging::fetch_all_changelogs SQL shows a full
table scan on the combined UNION ALL results (All 121,315 rows).

Limiting the UNION ALL data sets by applying the ORDER BY and LIMIT clauses to
the sub-queries improves the query performance to 0.20s.

Indexing changelog.ctype further improves query performance to 0.01s.